Fixed A non-numeric value encountered  when user creates a new topic in PHP 7.1
Fixed issue #28

http://stackoverflow.com/questions/42044127/warning-a-non-numeric-value-encountered

It seems that in PHP 7.1, a Warning will be emitted if a non-numeric
value is encountered.

Here is the relevant portion that pertains to the Warning notice you are
getting:

New E_WARNING and E_NOTICE errors have been introduced when invalid
strings are coerced using operators expecting numbers or their
assignment equivalents. An E_NOTICE is emitted when the string begins
with a numeric value but contains trailing non-numeric characters, and
an E_WARNING is emitted when the string does not contain a numeric
value.