Fixed incorrect URL for forum in Geeklog Search Results
For #87. Fixed Geeklog Search with forum results that used the topic id as the showtopic id (when it should have been the parent). Now proper page number is also used and jump anchor so when user clicks on the search result they are taken right to the result. (instead of just the start of the topic)

Forum search also now uses Geeklog's built in query highligh. This includes through the forum search and the Geeklog search.